moved package login in view for consistency
