+++ app/src/androidTest/java/ch/epfl/qedit/LoginActivityTest.java
@@ -10,7 +10,7 @@ import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
-import ch.epfl.qedit.ui.login.LoginActivity;
+import ch.epfl.qedit.view.login.LoginActivity;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
