+++ app/src/main/java/ch/epfl/qedit/view/MainActivity.java
@@ -3,11 +3,10 @@ package ch.epfl.qedit.view;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.EditText;
 
 import androidx.appcompat.app.AppCompatActivity;
 import ch.epfl.qedit.R;
-import ch.epfl.qedit.ui.login.LoginActivity;
+import ch.epfl.qedit.view.login.LoginActivity;
 
 public class MainActivity extends AppCompatActivity {
